Moved tests to test_functional, some renaming
